trabalhando com debugadores 2

+++ src/edu/sidnei/CalculaMedia.java
@@ -8,18 +8,18 @@ public class CalculaMedia {
         Scanner scan = new Scanner(System.in);
         String[] alunos = {"Aluno1", "Aluno2","Aluno3",};
 
-        int media = calculaMediaTurma(alunos, scan);
+        double media = calculaMediaTurma(alunos, scan);
 
         System.out.printf("Media da turma %d ",media);
 
 
     }
-    public static int calculaMediaTurma(String[] alunos, Scanner scanner){
-        int soma = 0;
+    public static double calculaMediaTurma(String[] alunos, Scanner scanner){
+        double soma = 0;
 
         for(String aluno : alunos){
             System.out.printf("Nota do aluno %s: ", alunos);
-            int nota = scanner.nextInt();
+            double nota = scanner.nextInt();
             soma = soma + nota;
         }
